+++ typedmatchers/interfaces.go
@@ -1,0 +1,5 @@
+package typedmatchers
+
+type binary interface {
+	[]byte | string
+}
