+++ typedmatchers/json_map.go
@@ -1,0 +1,5 @@
+package typedmatchers
+
+func JSONMap(m map[string]any) any {
+	return nil
+}
