+++ typedmatchers/json_map_test.go
@@ -1,0 +1,7 @@
+package typedmatchers
+
+// var
+
+// func TestJSONMap(t *testing.T) {
+// 	m := JSONMap()
+// }
